- add missing documentation - minor refactor
