+++ nsdb-http/src/main/scala/io/radicalbit/nsdb/web/routes/QueryApi.scala
@@ -32,7 +32,7 @@ import io.radicalbit.nsdb.protocol.MessageProtocol.Events._
 import io.radicalbit.nsdb.security.http.NSDBAuthProvider
 import io.radicalbit.nsdb.security.model.Metric
 import io.radicalbit.nsdb.sql.parser.SQLStatementParser
-import io.radicalbit.nsdb.web.BitSerializer.BitSerializer
+import io.radicalbit.nsdb.web.BitSerializer
 import io.radicalbit.nsdb.web.CustomSerializers
 import io.swagger.annotations._
 import javax.ws.rs.Path
